commit-dc

+++ Addition.java
@@ -1,3 +1,3 @@
-public static int addition(int a, int b) {
-    return a + b;
+public static double addition(int a, int b) {
+    return a + b ;
 }
